konsolitēts JS turpmākām aktivitātēm
