+++ templates/IIC.html
@@ -11,7 +11,7 @@
     <link rel='stylesheet' type='text/css' media='screen' href='css/main.css'>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV" crossorigin="anonymous"></script>
     <!--<script src='./main.js'></script>-->
-    <script src='../static/js/registracija.js'></script>
+    <script src='../static/js/main.js'></script>
 </head>
         
 <body>
