--- templates/fakeLandingPage.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset='utf-8'>
-    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
-    <title>Jautrais PC -- IIIIS</title>
-    <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <link rel='stylesheet' type='text/css' media='screen' href='/css/main.css'>
-    <script src='/js/main.js'></script>
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
-        integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
-</head>
-
-<body>
-    <div class="container">
-        <img src="./images/logo.png">
-        <h1>Izveidotās lapas</h1>
-        <a href="index.html">Sākums</a><br>
-        <a href="./views/skolenu_izvelne.html">Izvēles skats skolēnam</a><br>
-        <a href="./views/vecaku_pieteikums.html">Vecāku pieteikums</a><br>
-        <a href="IIC.html">Interešu izglītības centra pievienošanas skats</a><br>
-        <a href="./views/pulcina_info.html">Pulciņa pievienošanas skats</a><br>
-        <h1>Wireframe</h1>
-        <img src="https://docs.google.com/drawings/d/e/2PACX-1vRrVS-6eb6V9Gj068C5e5i7kGA7E3RBcQTSyXwBpyCDLkIxAwoeG4PKsiIv1Iyl71AMuH8idECdNlXg/pub?w=1070&amp;h=640">
-        <h1>Situācijas apraksts</h1>
-        <iframe
-            src="https://docs.google.com/document/d/e/2PACX-1vSUsgtGgp3UZI1eMB9g3CIOsSn7c1e6H-l1ONNb9QvV2SGeEdTY-Dzk0fgNtudsaPyutkCGNOWBH1hH/pub?embedded=true" style="width: 100%; height: 600px"></iframe>
-        <h1>Piecas problēmas</h1>
-        <h2>1. problēma</h2>
-<p>Bērns apslimis, kādam no vecākiem jāpaliek mājās, citu ģimenes locekļu palīdzība šodien nav pieejama.
-Taču ieplānota ļoti svarīga tikšanās, tādēļ uz divām stundām steidzami jāatrod auklīte. 
-Internetā var atrast daudz piedāvājumu, taču bieži auklītes negrib divu stundu dēļ tālu braukt vai  
-nepieciešamajā laikā ir aizņemtas.<br>
-Risinājums.<br>
-Mobilā lietotne “AUKLĪTIS”.<br>
-Auklīšu aģentūra uztur tīmekļa lapu, kurā klients var ievadīt savu atrašanās vietu un kartē ieraudzīt
-tuvumā esošo brīvo auklīšu atrašanās vietu un kontaktus.<br>
-Sīkāk - <a href="https://drive.google.com/file/d/1KcLjGDuE3HTJ-GhD3jPRyTxYsp2Cc9hD">https://drive.google.com/file/d/1KcLjGDuE3HTJ-GhD3jPRyTxYsp2Cc9hD</a>
-</p>
-
-<h2>2. problēma</h2>
-<p>Senioru bērniem bieži nākas zvanīt un atgādināt saviem vecākiem, ka pienācis laiks iedzert zāles.
-Taču bieži vien vajadzīgajā laikā piezvanīt nevar, jo jāpilda kādi neatliekami darba pienākumi.<br>
-Risinājums<br>
-Zāļu lietošanas atgādināšanas sistēma "Zaļā tabletīte".<br>
-Senioru aprūpētāji tajā iestata laikus, kad nepieciešami atgādinājumi.<br>
-Sistēma noteiktajā laikā zvana uz norādīto tālruņa numuru.<br>
-Sīkāk - <a href="https://docs.google.com/document/d/1mofmFIdUJ8csVD7Ev2DHKTIKpkvtoZkfkN-d2iq8jE8">https://docs.google.com/document/d/1mofmFIdUJ8csVD7Ev2DHKTIKpkvtoZkfkN-d2iq8jE8</a>
-</p>
-
-<h2>3. problēma </h2>
-<p>Bērni grib apmeklēt dažādus pulciņus, piedāvājums Rīgā ļoti plašs, taču informācija nav pieejama vienuviet. 
-Līdz ar to ir grūti atrast vēlamo nodarbību vēlamajā apkārtnē un laikā, it sevišķi, ja vēlas apmeklēt vairākus pulciņus.<br>
-Risinājums.<br>
-IIIS - interešu izglītības izvēļu informācijas sistēma.<br>
-Tīmekļa vietne, kurā interešu izglītības iestādes ievada informāciju par piedāvātajām nodarbībām, 
-bet skolēni un vecāki var filtrēt piedāvājumu pēc nodarbību veida, vietas un laika, un pieteikties uz izvēlēto pulciņu.<br>
-Sīkāk - <a href="https://docs.google.com/document/d/1wyqlTS-tAZXgj7CtUAqA6kbzvRIzw1ILZZmsmg6e96g">https://docs.google.com/document/d/1wyqlTS-tAZXgj7CtUAqA6kbzvRIzw1ILZZmsmg6e96g</a>
-</p>
-
-<h2>4. problēma</h2>
-<p>
-Skolas bibliotēkas datorsistēmā informāciju var ievadīt tikai manuāli, ar klaviatūras palīdzību.
-Tāpēc mācību gada sākumā un beigās, kad visiem skolēniem jāsaņem vai jānodod grāmatas, veidojas lielas rindas un bibliotekāres ir pārgurušas.<br>
-Risinājums<br>
-Esošās sistēmas papildinājums vai jauna sistēma, kas informācijas ievadam ļautu izmantot rokas skeneri.<br>
-</p>
-
-<h2>5. problēma</h2>
-<p>
-Sporta spēļu faniem patīk savā starpā sacensties par precīzākajām prognozēm.
-Līdz šim viņi to darīja, kopīgi skatoties attiecīgo spēli.
-Tācu Covids visu ir mainījis, tagad hokejs jāskatās mājās, katram pie sava TV. 
-Kā lai citi fani uzzina, kurš visprecīzāk prognizēja rezultatīvāko spēlētāju?<br>
-Risinājums<br>
-Sporta spēļu prognožu tīmekļa vietne.<br>
-Neatbildētais jautājums - vai un kur sporta spēļu laikā ir pieejami brīvie dati par iemesto grozu, rezultātu utml.<br>
-</p>
-
-    </div>
-</body>
-
-</html>
