--- templates/indexNotIndex.html
@@ -1,16 +1,0 @@
- <!DOCTYPE html>
- <html>
- <head>
-     <meta charset='utf-8'>
-     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
-     <title>IIIIS</title>
-     <meta name='viewport' content='width=device-width, initial-scale=1'>
-     <link rel='stylesheet' type='text/css' media='screen' href='main.css'>
-     <script src='main.js'></script>
- </head>
- <body>
-     <h1>Sveiks!</h1>
-     <p>Esmu veicis izmaiņas! jēēē!</p>
-     
- </body>
- </html>
