--- templates/pub_data.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Start(IT)</title>
-  </head>
-  <body>
-  <table id="pub_data">
-    <thead>
-      <tr>
-        <th>ID</th>
-        <th>Nosaukums</th>
-        <th>Tips</th>
-        <th>Apakštips</th>
-        <th>Skaits</th>
-        <th>Komentāri</th>
-      </tr>
-    </thead>
-    <tbody id="tbody">
-    </tbody>
-    </table>
-    <script src='js/fromAPI.js'></script>
-  </body>
-</html>
