+++ templates/pulcina_info.html
@@ -6,14 +6,14 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>IIIIS - pulciņa info</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-    <!-- <link rel='stylesheet' type='text/css' media='screen' href='/static/css/main.css'> -->
-    <script src='/static/js/sendPulcinaInfo.js'></script>
+    
+    <script src='../static/js/main.js'></script>
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
         integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
 </head>
 
-<body>
+<body onload="ieladetJomas()">
 
     <div class="container">
         <img src="/static/images/logo.png">
@@ -332,7 +332,7 @@
 
         </div>
     </div>
-    <script src='/static/js/ieladetKategorijas.js'></script>
+    <!-- <script src='/static/js/ieladetKategorijas.js'></script> -->
 </body>
 
 </html>
