+++ templates/pulcinu_saraksts.html
@@ -48,7 +48,7 @@
             </tbody>
 
         </table>
-        <script src='static/js/lasaPulcinuSarakstu.js'></script>
+        <script src='../static/js/main.js'></script>
         <button type="button" value="Jauns" class="btn btn-primary"
             onclick="location.href='/jauns_pulcins'">Reģistrēt jaunu</button>
