+++ templates/vecaku_pieteikums.html
@@ -5,15 +5,16 @@
     <meta http-equiv='X-UA-Compatible' content='IE=edge'>
     <title>Vecāku pieteikums</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
-	   <script src='../js/vecPietLasa.js'></script>
-	   <script src='../js/vecPietRegistree.js'></script>
+	<script src='../static/js/main.js'></script>
+	   <!-- <script src='../js/vecPietLasa.js'></script>
+	   <script src='../js/vecPietRegistree.js'></script> -->
 
 	<!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
         integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
 </head>
 
-<body onload="ieliktTekstu();">
+<body onload="lasitTekstu();">
 	<div class="container p-1 my-1 bg-success text-white">
         <img src="../images/logo.png" alt="logo">
